+++ packages/core/src/index.ts
@@ -5,3 +5,4 @@ export * from "./targets.js";
 export * from "./installer.js";
 export * from "./packs.js";
 export * from "./new-skill.js";
+export * from "./runtime.js";